This is the code after the reading competition. I have moved the calibration of the gyro to the robot init from init of auto and teleop. This removes the 5 second delays for both of them. Robot init occurs when the robot boots up so we need to keep the robot still during that time.
I have also adjusted the sensitivity of the x-axis on the joystick.

I have also removed the use of buttons that i originally programmed for testing auto. They should never have been enabled for our competition. We were lucky that they were not pressed, as they would make the robot unresponsive.

I have also added another option for auto, twice during the reading matches we would have apprechiated being able to only reach a defense and not cross if it if there was no defense that we were able to cross. there is now an option on the smart dashbaord to either cross or reach  a defense during auto.

Signed-off-by: dmatthewsshs <david.matthews@mosaicsystems.com>