Moved the broadcasting of smartDashboard data relating to the drivetrain to it's class.
Updated TurnAngleCmd to hopefully allow it to be called twice in a row without rebooting the robot -- during lunch (this commit), I was able to get the TurnAngleCmd to work on the first turn. the second time it would fail. I have not tested this current updated version. I plan to-do so tomorrow.